# Spectacular speedup in iter_ruptures