Finally, after ~40 pull requests in the project #7745 I am getting the expected speedup due to the vectorization. Here is the calculation by Guillaume:
```
# master
| calc_23919, maxmem=2.7 GB | time_sec  | memory_mb | counts  |
|---------------------------+-----------+-----------+---------|
| total classical           | 1_807     | 7.39453   | 95      |
| make_contexts             | 1_631     | 3.99609   | 30_645  |
| iter_ruptures             | 1_472     | 0.0       | 30_645  |
| ClassicalCalculator.run   | 299.1     | 296.9     | 1       |
| computing mean_std        | 82.9      | 0.0       | 166_031 |
| computing distances       | 14.2      | 0.0       | 225_274 |

# this branch
| calc_23918, maxmem=2.8 GB | time_sec  | memory_mb | counts  |
|---------------------------+-----------+-----------+---------|
| total classical           | 450.0     | 6.87891   | 95      |
| make_contexts             | 276.7     | 2.75781   | 30_645  |
| iter_ruptures             | 157.5     | 0.0       | 30_645  |
| computing mean_std        | 82.4      | 0.0       | 166_031 |
| ClassicalCalculator.run   | 81.8      | 297.3     | 1       |
| computing distances       | 46.3      | 0.0       | 30_645  |
```
The speedup in `iter_ruptures` is 9.4x and in `total classical` 4x. We can do still better. Notice that computing distances is not slower, we just starting including the time spent building the contexts under that umbrella (next it will be renamed as make_contexts).